Update data to no longer have celltype as an explicit variable. Instead, you can now specify ranks by a single celltype and also get ctspecific information for any number of celltypes without them being dependent. So, you could, for example, find cres active in one tissue and get the values for other tissues. With this, all ctspecific data is retrieved from the database rather than any one specific celltype. This probably won't affect performance much, if at all, but also allows caching of cres with a lower chance of requiring a new db call (since ctspecific data would have changed). However, this is unimplemented and would require more thought anyways, since ranks, order by, etc. are all db-controlled.
Also, removed element_type. It was unimplemented and would limit the user's control anyways.

Don't log exceptions to Raven if it's a UserError (this isn't a problem in the program, it's a problem in the user input). It might be interesting to see what errors users get (do they consistently use ct name instead of value, for example), but for now this is outside the scope of this feature.

Also, added in the option in globals to retrieve data from only a single cell type, rather than returning the whole list and filtering client side. This is useful for the search on screen, where we only want the ctinfo for the ct used in the search but nothing else.