Attempt to fix peak/cistrome intersection import. Now, cistrome intersection files mirror peak intersections. Notably:   1) Cistrome files now are generated into a folder labeled by run date   2) a new cistromeJobs.json.gz file is generated with metadata like peaks   3) cistromeIntersections.json.gz is now by runDate
This hasn't been tested, but the prior import files were broken. A (simplier)  alternative to mirroring the peak intersections by runDate would have been to change the ImportPeakIntersections class to be path-agnostic. However, being explicit about cistrome intersection dates and jobs is probably better.

Also note that the import file won't run correctly as of this commit unless the the cistromeIntersections.json.gz file was moved to a date folder and a cistromeJobs.json.gz file was generated.