Remove de_search and replace with de_search_window.
This now returns a window which includes all genes within the same 1 MB window centered on the queried gene. This window can then be used to query to genes and ccres to display for de. A few important changes have occured:
1) The window returned will most likely be larger in all cases since the window includes the full genes of all genes intersecting the minimum window versus only the de genes intersecting the window before. This can be fully compensated client side to only visualize the region around de genes. I don't foresee much additional overhead in terms of gene or ccre querying.
2) The center, type, and value of the ccREs overlapping the window are no longer calculated server side. Overall, this was a bit abritrary anyways. Now, these can be computed client-side since it is easy to query multiple celltype-specific data.
3) Most of the de code is much simpler (and more flexible).

NOTE: The de_search_window field is a bit abitrary and somewhat unnecessary. However, if it was removed, to get the same results either 2 or 3 client->server  requests would have to be made. Using all existing APIs: A gene could be queried. The client could expand the window based on a window. The client could then query this window for genes contained. The client could then slightly expand this window to fully enclose all genes. Finally the client could request fully all de data. If an 'expanded' field was added to ChromRange that allowed the calculating of the first expansion from gene coords to window, the first 2 requests could be merged. Alternatively, rather than the client-side de window including fully all genes, it could instead be altered to only display genes fully within the gene coords + half window. This plus the 'expanded' field allows nearly the same data, except in 1 request. Additionally, this makes conceptually querying an abritrary range of de data vs querying a gene nearly the same.